requirements.txt
modal
fastapi
uvicorn
llama-cpp-python[server]
pydantic
psutil
requests
